Ignore stupid flutter snakecase naming
