+++ lib/builders/bases/CustomCheckbox.dart
@@ -1,3 +1,4 @@
+// ignore_for_file: file_names
 import 'package:flutter/material.dart';
 
 class CustomCheckbox extends StatefulWidget {
