+++ lib/routes/data/AutonomousDataRoute.dart
@@ -5,12 +5,12 @@ import 'package:scouting_platform/routes/comments/CommentsRoute.dart';
 import 'package:scouting_platform/routes/data/fields/AutonomousRightRows/AutonomousRightRow2.dart';
 import 'package:scouting_platform/routes/data/fields/AutonomousRows.dart/AutonomousBottomReef.dart';
 import 'package:scouting_platform/routes/data/fields/AutonomousRows.dart/AutonomousMiddleReef.dart';
-import 'package:scouting_platform/routes/data/fields/AutonomouscCheckboxRows/AutonomousRow1.dart';
-import 'package:scouting_platform/routes/data/fields/AutonomouscCheckboxRows/AutonomousRow2.dart';
-import 'package:scouting_platform/routes/data/fields/AutonomouscCheckboxRows/AutonomousRow3.dart';
-import 'package:scouting_platform/routes/data/fields/AutonomouscCheckboxRows/AutonomousRow4.dart';
-import 'package:scouting_platform/routes/data/fields/AutonomouscCheckboxRows/AutonomousRow5.dart';
-import 'package:scouting_platform/routes/data/fields/AutonomouscCheckboxRows/AutonomousRow6.dart';
+import 'package:scouting_platform/routes/data/fields/AutonomousCheckboxRows/AutonomousRow1.dart';
+import 'package:scouting_platform/routes/data/fields/AutonomousCheckboxRows/AutonomousRow2.dart';
+import 'package:scouting_platform/routes/data/fields/AutonomousCheckboxRows/AutonomousRow3.dart';
+import 'package:scouting_platform/routes/data/fields/AutonomousCheckboxRows/AutonomousRow4.dart';
+import 'package:scouting_platform/routes/data/fields/AutonomousCheckboxRows/AutonomousRow5.dart';
+import 'package:scouting_platform/routes/data/fields/AutonomousCheckboxRows/AutonomousRow6.dart';
 import 'package:scouting_platform/routes/data/fields/AutonomousRightRows/AutonomousRightRow1.dart';
 import 'package:scouting_platform/routes/data/fields/AutonomousRightRows/AutonomousRightRow3.dart';
 import 'package:scouting_platform/routes/data/labels/AutonomousRightLabels/AutonomousRightLabel1.dart';
