+++ lib/routes/data/fields/AutonomousCheckboxRows/AutonomousRow3.dart
@@ -1,17 +1,18 @@
+// ignore_for_file: file_names
 import 'package:flutter/material.dart';
 import 'package:scouting_platform/utils/data/values/AutonomousValues.dart';
 import 'package:scouting_platform/builders/bases/CustomCheckbox.dart';
 
-class AutonomousRow1 extends StatefulWidget {
-  const AutonomousRow1({
+class AutonomousRow3 extends StatefulWidget {
+  const AutonomousRow3({
     super.key,
   });
 
   @override
-  State<AutonomousRow1> createState() => _AutonomousRow1State();
+  State<AutonomousRow3> createState() => _AutonomousRow3State();
 }
 
-class _AutonomousRow1State extends State<AutonomousRow1> {
+class _AutonomousRow3State extends State<AutonomousRow3> {
   /// Increments an integer in a controller's value by one
   void incrementNumber(TextEditingController controller) {
     if (!mounted) return;
@@ -44,46 +45,46 @@ class _AutonomousRow1State extends State<AutonomousRow1> {
         children: [
           // A
           CustomCheckbox(
-              controller: AutonomousValues.l4A,
+              controller: AutonomousValues.l2A,
               backgroundColor: Colors.grey.shade800,
               checkColor: Colors.white,
               labelColor: Colors.white,
               margin: const EdgeInsets.only(right: 18.0)),
           // L
           CustomCheckbox(
-              controller: AutonomousValues.l4L,
+              controller: AutonomousValues.l2L,
               backgroundColor: Colors.grey.shade800,
               checkColor: Colors.white,
               labelColor: Colors.white,
               margin: const EdgeInsets.only(right: 35.0)),
           // K
           CustomCheckbox(
-              controller: AutonomousValues.l4K,
+              controller: AutonomousValues.l2K,
               backgroundColor: Colors.grey.shade800,
               checkColor: Colors.white,
               labelColor: Colors.white,
               margin: const EdgeInsets.only(right: 35.0)),
           // J
           CustomCheckbox(
-              controller: AutonomousValues.l4J,
+              controller: AutonomousValues.l2J,
               backgroundColor: Colors.grey.shade800,
               checkColor: Colors.white,
               labelColor: Colors.white,
               margin: const EdgeInsets.only(right: 30.0)),
           // I
           CustomCheckbox(
-              controller: AutonomousValues.l4I,
+              controller: AutonomousValues.l2I,
               backgroundColor: Colors.grey.shade800,
               checkColor: Colors.white,
               labelColor: Colors.white,
               margin: const EdgeInsets.only(right: 25.0)),
           // H
           CustomCheckbox(
-              controller: AutonomousValues.l4H,
+              controller: AutonomousValues.l2H,
               backgroundColor: Colors.grey.shade800,
               checkColor: Colors.white,
               labelColor: Colors.white,
-              margin: const EdgeInsets.all(00.0)),
+              margin: const EdgeInsets.only(right: 0.0)),
         ],
       ),
     );
