+++ lib/routes/data/fields/AutonomousRightRows/AutonomousRightRow3.dart
@@ -1,3 +1,4 @@
+// ignore_for_file: file_names
 import 'package:flutter/material.dart';
 import 'package:scouting_platform/builders/bases/CounterNumberField.dart';
 import 'package:scouting_platform/utils/data/values/AutonomousValues.dart';
